# openapi.yaml
openapi: 3.1.0
info:
  title: its-laika/NanoBlog
  description: |-
    _a simplistic blogging system_

    NanoBlog is a blogging system on file basis, therefore it doesn't use a database or cookies.
  license:
    name: MIT
    url: https://github.com/its-laika/NanoBlog/blob/main/LICENSE
  version: 0.2.0
externalDocs:
  description: Github repository
  url: https://github.com/its-laika/NanoBlog
servers:
  - url: http://localhost:5023/api/
tags:
  - name: assets
    description: Control files that can be included into the blog
  - name: posts
    description: Control the posts of the blog
  - name: export
    description: Control exporting stashed changes
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    MimeTypeError:
      description: |-
        MIME type error (e.g. MIME type is not identifiable or MIME types differ when trying to update existing asset)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    CreatedAt:
      description: A resource was created successfully
      headers:
        Location:
          schema:
            type: string
          description: URL of the new asset. Given file name will be overridden
    FileList:
      description: Responded with a list of existing file names
      content:
        application/json:
          schema:
            type: array
            items:
              type: string
              description: File name
              examples:
                - "638075963053552470-0747b6f6-c677-49fb-bbde-53addf7b3b71.png"
    PostExcerpts:
      description: Responded with a list of post excerpts
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              description: Post excerpt
              properties:
                name:
                  type: string
                  description: Post file name
                  examples:
                    - "638075963053552470-0747b6f6-c677-49fb-bbde-53addf7b3b71.txt"
                excerpt:
                  type: string
                  description: Excerpt of the post
                  examples:
                    - "<p>Lorem ipsum</p>"
  schemas:
    Error:
      type: object
      properties:
        type:
          type: string
          examples:
            - "https://www.rfc-editor.org/rfc/rfc2324#section-2.3.2"
        title:
          type: string
          examples:
            - "I'm a teapot"
        status:
          type: number
          examples:
            - 418
        traceId:
          type: string
          examples:
            - "00-420b7c937b9403217d4b8e97be0bef05-024ecd9aa651ee47-00"
      required:
        - type
        - title
        - status
        - traceId
security:
  - bearerAuth: []
paths:
  /assets:
    get:
      tags:
        - assets
      summary: Gets list of assets
      description: Gets list of all stored assets names
      responses:
        "200":
          $ref: "#/components/responses/FileList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - assets
      summary: Stores new asset
      description: Stores new asset by given file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: File to be stored as form data string
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          $ref: "#/components/responses/CreatedAt"
        "400":
          $ref: "#/components/responses/MimeTypeError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /assets/{fileName}:
    get:
      parameters:
        - name: fileName
          in: path
          description: File name of asset
          required: true
          schema:
            type: string
      tags:
        - assets
      summary: Returns assets file content
      description: Returns file content by asset file name
      responses:
        "200":
          description: Stored asset file
          content:
            image/*:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      parameters:
        - name: fileName
          in: path
          description: File name of asset
          required: true
          schema:
            type: string
      tags:
        - assets
      summary: Updates an asset
      description: Updates an existing asset by asset file name
      requestBody:
        description: New file to be stored
        content:
          image/*:
            schema:
              type: string
              format: binary
        required: true
      responses:
        "201":
          description: Successful operation
        "400":
          $ref: "#/components/responses/MimeTypeError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - assets
      summary: Deletes an asset
      description: Deletes an asset by file name
      parameters:
        - name: fileName
          in: path
          description: File name of asset
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Successful operation
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /posts/excerpts:
    get:
      parameters:
        - name: length
          in: query
          description: Max length of post excerpt
          required: false
          schema:
            type: number
      tags:
        - posts
      summary: Returns excerpts of all posts
      description: Returns excerpts of all posts with post file name
      responses:
        "200":
          $ref: "#/components/responses/PostExcerpts"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /posts:
    get:
      tags:
        - posts
      summary: Gets list of posts
      description: Gets list of all stored posts names
      responses:
        "200":
          $ref: "#/components/responses/FileList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - posts
      summary: Stores new post
      description: Stores new post by given content
      requestBody:
        description: HTML (or plain text) post content
        required: true
        content:
          text/html:
            schema:
              type: string
      responses:
        "201":
          $ref: "#/components/responses/CreatedAt"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /posts/{fileName}:
    get:
      parameters:
        - name: fileName
          in: path
          description: File name of post
          required: true
          schema:
            type: string
      tags:
        - posts
      summary: Returns post content
      description: Returns content by post file name
      responses:
        "200":
          description: Stored post content
          content:
            text/html:
              schema:
                type: string
                examples:
                  - "<h1>title</h1>this is my content"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      parameters:
        - name: fileName
          in: path
          description: File name of post
          required: true
          schema:
            type: string
      tags:
        - posts
      summary: Updates a post
      description: Updates an existing post by new content
      requestBody:
        description: New post content
        content:
          text/html:
            schema:
              type: string
        required: true
      responses:
        "201":
          description: Successful operation
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - posts
      summary: Deletes a post
      description: Deletes a post by file name
      parameters:
        - name: fileName
          in: path
          description: File name of post
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Successful operation
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /export:
    post:
      tags:
        - export
      summary: Exports stashed changes
      description: Exports stashed changes to directory and synchronizes assets
      responses:
        "201":
          description: Successful operation
        "401":
          $ref: "#/components/responses/Unauthorized"